notification remove and change logoout
