+++ src/components/Navbar.tsx
@@ -1,10 +1,11 @@
 'use client'
 
 import { useState } from 'react';
-import { Bell, Menu, X } from 'lucide-react';
+import { Bell, Menu, X, LogOut } from 'lucide-react';
 import Image from 'next/image';
 import logo from '../../public/logo.png';
 import admin from '../../public/admin.jpg';
+
 
 interface Notification {
   id: number;
@@ -26,10 +27,27 @@ interface AdminProfileProps {
   role?: string;
 }
 
-const Navbar = ( { role, userid }: { role: string; userid: string }) => {
+const Navbar = ({ role, userid }: { role: string; userid: string }) => {
   console.log(userid)
   const [showNotifications, setShowNotifications] = useState<boolean>(false);
   const [isMobileMenuOpen, setIsMobileMenuOpen] = useState<boolean>(false);
+  const handleLogout = async () => {
+    try {
+      const response = await fetch("/api/Auth/logout", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+      });
+      if (response.ok) {
+        window.location.href = "/";
+      } else {
+        console.error("Logout failed");
+      }
+    } catch (error) {
+      console.error("Error during logout:", error);
+    }
+  };
 
   // Demo notifications
   const notifications: Notification[] = [
@@ -97,24 +115,39 @@ const Navbar = ( { role, userid }: { role: string; userid: string }) => {
 
           {/* Desktop Menu */}
           <div className="hidden lg:flex items-center gap-6">
-            <NotificationButton 
-              notifications={notifications} 
-              showNotifications={showNotifications}
-              toggleNotifications={toggleNotifications}
-            />
-            <AdminProfile role={role ?? ''}/>
+            <button
+              onClick={handleLogout}
+              className={`
+            flex items-center h-12 px-4 text-red-500 hover:bg-gray-50
+          `}
+            >
+              <LogOut className="h-5 w-5 min-w-[20px]" />
+              <span className="hidden lg:block ml-4 text-sm font-medium">Logout</span>
+              <div className="lg:hidden absolute left-16 px-3 py-2 bg-gray-800 text-white text-sm rounded opacity-0 invisible group-hover:opacity-100 group-hover:visible transition-all duration-200 whitespace-nowrap z-50">
+                Logout
+              </div>
+            </button>
+
+            <AdminProfile role={role ?? ''} />
           </div>
 
           {/* Mobile Menu */}
           {isMobileMenuOpen && (
             <div className="lg:hidden absolute top-16 left-0 right-0 bg-white border-t border-gray-100 shadow-lg">
               <div className="p-4 space-y-4">
-                <NotificationButton 
-                  notifications={notifications} 
-                  showNotifications={showNotifications}
-                  toggleNotifications={toggleNotifications}
-                  isMobile={true}
-                />
+                <button
+                  onClick={handleLogout}
+                  className={`
+            flex items-center h-12 px-4 text-red-500 hover:bg-gray-50
+          `}
+                >
+                  <LogOut className="h-5 w-5 min-w-[20px]" />
+                  <span className="hidden lg:block ml-4 text-sm font-medium">Logout</span>
+                  <div className="lg:hidden absolute left-16 px-3 py-2 bg-gray-800 text-white text-sm rounded opacity-0 invisible group-hover:opacity-100 group-hover:visible transition-all duration-200 whitespace-nowrap z-50">
+                    Logout
+                  </div>
+                </button>
+
                 <div className="border-t border-gray-100 pt-4">
                   <AdminProfile isMobile={true} />
                 </div>
@@ -127,66 +160,12 @@ const Navbar = ( { role, userid }: { role: string; userid: string }) => {
   );
 };
 
-// Notification Button Component
-const NotificationButton: React.FC<NotificationButtonProps> = ({ 
-  notifications, 
-  showNotifications, 
-  toggleNotifications, 
-  isMobile = false 
-}) => {
-  const unreadCount = notifications.filter(n => !n.isRead).length;
-  
-  return (
-    <div className={`relative ${isMobile ? 'w-full' : ''}`}>
-      <button 
-        onClick={toggleNotifications}
-        className={`relative p-2 rounded-full hover:bg-purple-50 transition-colors
-          ${isMobile ? 'w-full flex items-center gap-3 px-4' : ''}`}
-        type="button"
-        aria-label="Toggle notifications"
-      >
-        <Bell className="h-5 w-5 text-black" />
-        <span className="absolute top-0 right-0 bg-red-500 text-white text-xs rounded-full h-5 w-5 flex items-center justify-center">
-          {unreadCount}
-        </span>
-        {isMobile && <span className="text-gray-700">Notifications</span>}
-      </button>
 
-      {showNotifications && (
-        <div className={`
-        bg-[#6800E9] rounded-lg shadow-lg border border-gray-100 overflow-hidden 
-          ${isMobile ? 'w-full mt-2' : 'absolute right-0 mt-2 w-80'}
-        `}>
-          <div className="p-4 border-b border-gray-100">
-            <h3 className="text-lg font-semibold text-white text-center">Notifications</h3>
-          </div>
-          <div className="max-h-96 overflow-y-auto">
-            {notifications.map((notification) => (
-              <div
-                key={notification.id}
-                className={`p-4 border-b border-gray-100 hover:bg-gray-50 transition-colors
-                  ${!notification.isRead ? 'bg-purple-50' : ''}`}
-              >
-                <div className="flex justify-between items-start">
-                  <h4 className="text-sm font-semibold text-gray-800">{notification.title}</h4>
-                  <span className="text-xs text-gray-500">{notification.time}</span>
-                 
-                </div>
-                <p className="text-sm text-gray-600 mt-1">{notification.message}</p>
-                <hr />
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
-    </div>
-  );
-};
 
 // Admin Profile Component
 const capitalize = (str: string) => str.charAt(0).toUpperCase() + str.slice(1);
-const AdminProfile: React.FC<AdminProfileProps> = ({ role,isMobile = false}) => (
-  
+const AdminProfile: React.FC<AdminProfileProps> = ({ role, isMobile = false }) => (
+
   <div className={`flex items-center ${isMobile ? 'w-full' : 'gap-3'}`}>
     <div className={`${isMobile ? 'flex-1' : 'text-right'}`}>
       <p className="text-sm font-medium text-gray-700">{role ? capitalize(role) : ''}</p>
