+++ src/components/Sidebar.tsx
@@ -14,23 +14,7 @@ import {
 
 export default function Sidebar({ role, userid }: { role: string; userid: string }) {
 console.log(userid)
-  const handleLogout = async () => {
-    try {
-      const response = await fetch("/api/Auth/logout", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-      if (response.ok) {
-        window.location.href = "/";
-      } else {
-        console.error("Logout failed");
-      }
-    } catch (error) {
-      console.error("Error during logout:", error);
-    }
-  };
+  
 
 
   const pathname = usePathname()
@@ -94,18 +78,7 @@ console.log(userid)
         </div>
 
         {/* Logout Button */}
-        <button
-          onClick={handleLogout}
-          className={`
-            flex items-center h-12 px-4 text-red-500 hover:bg-gray-50
-          `}
-        >
-          <LogOut className="h-5 w-5 min-w-[20px]" />
-          <span className="hidden lg:block ml-4 text-sm font-medium">Logout</span>
-          <div className="lg:hidden absolute left-16 px-3 py-2 bg-gray-800 text-white text-sm rounded opacity-0 invisible group-hover:opacity-100 group-hover:visible transition-all duration-200 whitespace-nowrap z-50">
-            Logout
-          </div>
-        </button>
+       
       </nav>
     </aside>
   )
